docs: update API docs for renderer state refactoring (June 2026)
- Update README.md, docs/CHANGELOG.md, docs/ORIENTATION.md with R3 section
- Document new aggregates: RendererFrameHistory, RendererDirtyTelemetry,
  RendererLayerHistory, LayerBBoxState, RendererBufferRing,
  TransformScratchBuffer, CompiledGraphCache
- Document breaking changes: TextAnchor enum, SceneBuilder explicit include,
  project_layer_2_5d default parameter, software_renderer_internal.hpp removal
- Fix tests: add missing includes, update TextAnchor usage, fix type mismatch
- Fix project_layer_2_5d Mat4 overload: add diagnostics_enabled default

+++ tests/core/test_render_counters.cpp
@@ -253,7 +253,7 @@ TEST_CASE("RenderCounters: thread_local is faster than atomic on N threads") {
     // threads, otherwise the refactor is a regression.  In practice
     // it is 2-5x faster on a 4+ core machine.
     MESSAGE("atomic_ns=", atomic_ns, " tls_ns=", tls_ns,
-            " speedup_x=", static_cast<double>(atomic_ns) / std::max(tls_ns, int64_t{1}));
+            " speedup_x=", static_cast<double>(atomic_ns) / std::max(tls_ns, double{1.0}));
     CHECK(tls_ns <= atomic_ns * 2);
 }
